pyproject.toml
[project]
name = "pythinfer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Robert Muil", email = "robertmuil@gmail.com" }]
requires-python = ">=3.12"
dependencies = ["pyyaml>=6.0.3", "rdflib>=7.4.0", "typer>=0.20.0"]

[project.scripts]
pythinfer = "pythinfer.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["COM812"]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "pandas>=2.3.3",
]
